ci: add dockerfile to package into container
This adds a dockerfile to the repo which is responsible for packaging
the application into a container. It utilizes a multi-stage build in
order to first install the necesary dependencies and build the
application and in the final build simply copy the binary from the first
stage. It also adds a docker compose file in order to make development
easier.