ci: add code quality job
This uses the template given by GitLab to run code quality checks. This
job runs on merge requests, master and tags, but doesn't run in branches
that are not used in merge requests. This also exports a code quality
report so that gitlab can show the status in the merge request view